refactor: 디렉터리명 복수로 수정
