+++ src/App.js
@@ -1,9 +1,9 @@
-import LottoController from "./controller/LottoController.js";
-import LottoPurchaseController from "./controller/LottoPurchaseController.js";
-import LottoResultController from "./controller/LottoResultController.js";
-import WinningNumberController from "./controller/WinningNumberController.js";
-import InputView from "./view/InputView.js";
-import OutputView from "./view/OutputView.js";
+import LottoController from "./controllers/LottoController.js";
+import LottoPurchaseController from "./controllers/LottoPurchaseController.js";
+import LottoResultController from "./controllers/LottoResultController.js";
+import WinningNumberController from "./controllers/WinningNumberController.js";
+import InputView from "./views/InputView.js";
+import OutputView from "./views/OutputView.js";
 
 class App {
   async run() {
